Se crearon los metodos get y post para todas las entidades

+++ backend/src/main/java/com/arjusven/backend/repository/TicketRepository.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface TicketRepository extends JpaRepository<Tickets, Integer> {
+public interface TicketRepository extends JpaRepository<Tickets, Long> {
